Move code into `lib/` subdirectory.
