+++ .eslintrc.js
@@ -18,10 +18,7 @@ module.exports = {
         '.prettierrc.js',
         '.template-lintrc.js',
         'ember-cli-build.js',
-        'ember-addon-main.js',
-        'index.js',
-        'addDependencyTracker.js',
-        'utils.js',
+        'lib/**/*.js',
         'testem.js',
         'blueprints/*/index.js',
         'config/**/*.js',
