+++ lib/addDependencyTracker.js
