+++ lib/ember-addon-main.js
@@ -6,7 +6,7 @@ const addDependencyTracker = require('./addDependencyTracker');
 const hashForDep = require('hash-for-dep');
 
 module.exports = {
-  name: require('./package').name,
+  name: require('../package').name,
 
   parentRegistry: null,
 
@@ -45,7 +45,7 @@ module.exports = {
       _addon: this,
       toTree(tree) {
         let htmlbarsOptions = this._addon.htmlbarsOptions();
-        let TemplateCompiler = require('./index');
+        let TemplateCompiler = require('./template-compiler-plugin');
         return new TemplateCompiler(tree, htmlbarsOptions);
       },
