+++ lib/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  TemplateCompilerPlugin: require('./template-compiler-plugin'),
+};
