+++ lib/template-compiler-plugin.js
